While checking and constructing CSCCFEBTimeSlice objects in CSCCFEBData constructor there are 3 posibilities:
 1. B-word for TimeSlice is set and this particular slice is marked as FALSE,
 2. CSCCFEBTimeSlice::check() returns 1 and it is marked as TRUE,
 3. CSCCFEBTimeSlice::check() returns 0 and it is just skipped!

The 3rd cases troubles because in the current version of unpacker it is not being marked as FALSE.

Afterwards, while trying to access a timeslice that has higher (!) index number than the skipped one
we get into undefined behaviour. Because slice placeholder in vector was not initialized but just allocated.
So if c++ has initialized FALSE for status - we get nothing. Otherwise - we can get SEGFAIL because
slice position in the buffer was not initialized as well. Additionally, this screwds up the indexing of
TimeSlices as well.